Rename invoice to result
Currently, indexer has two invoice types. One is the invoice returned by
CodeChain, and the other is the invoice that indexer makes. The first
invoice is a boolean value that means the success of the transaction.
The second is an object `{ success: boolean, errorHint?: string }`.
This patch renames the first invoice to result since CodeChain renames
it, while the second invoice is left. So the object
`{ success: boolean, errorHint?: string }` becomes the only invoice type
from now.