Do not set the timeout per test
The most time consuming part of the tests is synchronizing the db with
the chain, and it depends on the chain. In other words, the test needs
a longer time when the chain has more data. So set the timeout per test
is meaningless in the most case.

This patch makes `yarn test` sets a global timeout instead of setting
the timeout per tests.