Optimize the API to get transactions
The previous query joins all rows of tables to calculate offset even
though it needs only itemsPerPage rows.

This patch splits the query to get transactions. The first query gets
the transaction hash that the second query selects. It helps the second
query not to scan all the table of the rows of included tables.