Add "transactionIndex" column to the UTXOs table
The indexer should sort UTXOs after queries. Sorting results by the
`id` field is not a good option. Let's think about a UTXO that is
created by a pending transaction. It's ` id` field could be lower than
UTXOs created by mined transactions.

Using [`blockNumber`, `transactionIndex`, and
`transactionOutputIndex`] makes a total order.

This commit adds the `transactionIndex` column to the `UTXOs` table to
use it as an index.