Add AggsUTXOs table
Users who use a wallet want to know which amount of assets they
have. The indexer was calculating the data using an aggregation
query. The aggregation query was slow as the aggregation query should
scan the UTXOs.

To make the query faster, we can make a table that has pre-calculated
asset amounts.

This commit creates the AggsUTXOs table. This table contains the
pre-calculated the amounts of assets. For the data consistency, we are
using the PostgreSQL's trigger feature to update the quantities.