Use AggsUTXOs table in the AggsUTXO query
Users who use a wallet want to know which amount of assets they
have. The indexer was calculating the data using an aggregation
query. The aggregation query was slow as the aggregation query should
scan the UTXOs.

To make the query faster, we can make a table that has pre-calculated
asset amounts.

This commit makes the UTXO aggregation API use the pre-calculated table.